requirements.txt
Django==4.2.2
django-allauth==0.54.0
django-jazzmin==2.6.0
gunicorn==20.1.0
mysqlclient==2.2.0
Pillow==9.5.0
python-dotenv==1.0.0
requests==2.31.0
whitenoise==6.2.0